third_party/blink/web_tests/svg/dom/SVGSVGElement-intersection-enclosure-expected.txt

Verify SVGSVGElement's getIntersectionList() and getEnclosureList() behavior.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



rect: [50 50 150 150]
referenceElement: null
PASS svg.getEnclosureList(rect, reference_element).length is 100
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [55 55 145 145]
referenceElement: null
PASS svg.getEnclosureList(rect, reference_element).length is 64
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [50 50 150 150]
referenceElement: container
PASS svg.getEnclosureList(rect, reference_element).length is 100
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [55 55 145 145]
referenceElement: container
PASS svg.getEnclosureList(rect, reference_element).length is 64
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [50 50 150 150]
referenceElement: subcontainer
PASS svg.getEnclosureList(rect, reference_element).length is 50
PASS svg.getIntersectionList(rect, reference_element).length is 50


rect: [55 55 145 145]
referenceElement: subcontainer
PASS svg.getEnclosureList(rect, reference_element).length is 32
PASS svg.getIntersectionList(rect, reference_element).length is 50


Zooming should not affect the results.


rect: [50 50 150 150]
referenceElement: null
PASS svg.getEnclosureList(rect, reference_element).length is 100
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [55 55 145 145]
referenceElement: null
PASS svg.getEnclosureList(rect, reference_element).length is 64
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [50 50 150 150]
referenceElement: container
PASS svg.getEnclosureList(rect, reference_element).length is 100
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [55 55 145 145]
referenceElement: container
PASS svg.getEnclosureList(rect, reference_element).length is 64
PASS svg.getIntersectionList(rect, reference_element).length is 100


rect: [50 50 150 150]
referenceElement: subcontainer
PASS svg.getEnclosureList(rect, reference_element).length is 50
PASS svg.getIntersectionList(rect, reference_element).length is 50


rect: [55 55 145 145]
referenceElement: subcontainer
PASS svg.getEnclosureList(rect, reference_element).length is 32
PASS svg.getIntersectionList(rect, reference_element).length is 50


But a viewbox transform should.


rect: [50 50 150 150]
referenceElement: null
PASS svg.getEnclosureList(rect, reference_element).length is 36
PASS svg.getIntersectionList(rect, reference_element).length is 49


rect: [55 55 145 145]
referenceElement: null
PASS svg.getEnclosureList(rect, reference_element).length is 25
PASS svg.getIntersectionList(rect, reference_element).length is 36


rect: [50 50 150 150]
referenceElement: container
PASS svg.getEnclosureList(rect, reference_element).length is 36
PASS svg.getIntersectionList(rect, reference_element).length is 49


rect: [55 55 145 145]
referenceElement: container
PASS svg.getEnclosureList(rect, reference_element).length is 25
PASS svg.getIntersectionList(rect, reference_element).length is 36


rect: [50 50 150 150]
referenceElement: subcontainer
PASS svg.getEnclosureList(rect, reference_element).length is 6
PASS svg.getIntersectionList(rect, reference_element).length is 14


rect: [55 55 145 145]
referenceElement: subcontainer
PASS svg.getEnclosureList(rect, reference_element).length is 0
PASS svg.getIntersectionList(rect, reference_element).length is 6


PASS successfullyParsed is true

TEST COMPLETE
